move sqldelight generated files to dedicated package
