+++ src/commonMain/kotlin/fr/acinq/phoenixd/db/SqliteChannelsDb.kt
@@ -21,6 +21,7 @@ import fr.acinq.bitcoin.ByteVector32
 import fr.acinq.lightning.CltvExpiry
 import fr.acinq.lightning.channel.states.PersistedChannelState
 import fr.acinq.lightning.db.ChannelsDb
+import fr.acinq.phoenixd.db.sqldelight.PhoenixDatabase
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
